+++ controllers/userController.js
@@ -13,8 +13,6 @@ export const createUser = async (req, res) => {
     if (existingUser.length > 0) return res.status(400).json("User already exist.")
     if (user.password != user.confirmPassword) return res.status(400).json("Password must match")
     // hash the password before insert it into database
-    // const hashedPassword = bcrypt.hashSync(user.password, 12);
-    // change salt to a random number between 20-40, instead of static 12
     const hashedPassword = bcrypt.hashSync(user.password, Math.round((Math.random() + 1) * 10));
     const now = new Date();
     data = await userModel.createUser(
@@ -48,7 +46,11 @@ export const getAllUsers = async (req, res) => {
 export const getUserById = async (req, res) => {
   try {
     const data = await userModel.getUserById(req.params.id);
-    res.status(200).json(data)
+    if (data[0].userId == req.user.userId || req.user.isAdmin == 1) {
+      res.status(200).json(data)
+    } else {
+      res.status(403).json("You are not authorized")
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json("getUserById query error")
@@ -77,6 +79,7 @@ export const getUserBySearch = async (req, res) => {
 // update user
 export const updateUser = async (req, res) => {
   const user = req.body;
+  if (user.userId !== req.user.userId && req.user.isAdmin != 1) return res.status(403).json("You are not authorized")
   let hashedPassword = user.password;
   !user.password.startsWith("$2b$") && (hashedPassword = bcrypt.hashSync(user.password, 12))
   try {
@@ -98,8 +101,9 @@ export const updateUser = async (req, res) => {
 }
 //delete user
 export const deleteUser = async (req, res) => {
+  let userId = req.params.id
+  if (req.user.isAdmin != 1) return res.status(403).json("You are not authorized")
   try {
-    let userId = req.params.id
     const data = await userModel.deleteUser(userId)
     await deleteCart(userId)
     await emptyCart(userId)
@@ -121,13 +125,22 @@ export const login = async (req, res) => {
     // if the email and password both correct, create a jwt token for the user, and send it to the user
     const token = jwt.sign(
       {
-        id: user[0].userId,
+        userId: user[0].userId,
         isAdmin: user[0].isAdmin
       },
       process.env.JWT_SECRET,
       { expiresIn: "2h" }
     );
-    res.status(200).json({ ...user[0], token, password: "" })
+    res.cookie(
+      'token',
+      token,
+      {
+        sameSite: 'None',
+        expires: new Date(new Date().getTime() + 2 * 60 * 60 * 1000),
+        httpOnly: true,
+        secure: true
+      })
+    res.status(200).json({ ...user[0], password: "" })
   } catch (error) {
     console.log(error);
     res.status(500).json("Login query error")
